Moved windows glob() over to dev/unistd.h
Added header guard and changed long->int64_t in dataloader.h
